BUGFIX: SD-829 SD-842  - SD-829: ignore ratio  - SD-827: fix js error when add filter  - TRIVIAL: refactoring geoChart storybook data, fix UI Legend
