+++ src/helpers/tests/geoChart/data.spec.ts
@@ -267,7 +267,7 @@ describe("geoChartData", () => {
                     "Rhode Island",
                     "Oregon",
                     "New York",
-                    "New Hempshire",
+                    "New Hampshire",
                     "Nebraska",
                     "Kansas",
                     "Mississippi",
@@ -285,6 +285,7 @@ describe("geoChartData", () => {
                     "Hawaii",
                     "Alaska",
                     "Tennessee",
+                    "Virginia",
                     "New Jersey",
                     "Kentucky",
                     "North Dakota",
@@ -295,20 +296,19 @@ describe("geoChartData", () => {
                     "Utah",
                     "Colorado",
                     "Ohio",
-                    "Utah",
+                    "Alabama",
                     "Iowa",
                     "New Mexico",
                     "South Carolina",
                     "Pennsylvania",
                     "Arizona",
-                    "Marryland",
+                    "Maryland",
                     "Massachusetts",
                     "California",
                     "Idaho",
                     "Wyoming",
                     "North Carolina",
                     "Louisiana",
-                    "Waikapu",
                 ],
                 index: 2,
                 name: "State",
@@ -448,7 +448,7 @@ describe("geoChartData", () => {
                     "Rhode Island",
                     "Oregon",
                     "New York",
-                    "New Hempshire",
+                    "New Hampshire",
                     "Nebraska",
                     "Kansas",
                     "Mississippi",
@@ -466,6 +466,7 @@ describe("geoChartData", () => {
                     "Hawaii",
                     "Alaska",
                     "Tennessee",
+                    "Virginia",
                     "New Jersey",
                     "Kentucky",
                     "North Dakota",
@@ -476,20 +477,19 @@ describe("geoChartData", () => {
                     "Utah",
                     "Colorado",
                     "Ohio",
-                    "Utah",
+                    "Alabama",
                     "Iowa",
                     "New Mexico",
                     "South Carolina",
                     "Pennsylvania",
                     "Arizona",
-                    "Marryland",
+                    "Maryland",
                     "Massachusetts",
                     "California",
                     "Idaho",
                     "Wyoming",
                     "North Carolina",
                     "Louisiana",
-                    "Waikapu",
                 ],
                 index: 1,
                 name: "State",
