+++ stories/core_components/GeoPushpinChart.tsx
@@ -155,6 +155,7 @@ storiesOf("Core components/GeoPushpinChart", module)
                     "/gdc/md/storybook/obj/694/elements?id=1903",
                     "/gdc/md/storybook/obj/694/elements?id=1870",
                     "/gdc/md/storybook/obj/694/elements?id=1895",
+                    "/gdc/md/storybook/obj/694/elements?id=1844",
                 ],
             },
         };
