+++ stories/test_data/geo_chart/geo_chart_with_location_size_color_segment_tooltip_filters_result.js
@@ -1,12 +1,23 @@
 // (C) 2020 GoodData Corporation
+const {
+    SIZE_AFM_DATA,
+    COLOR_AFM_DATA,
+    getLocationAFMData,
+    getSegmentAFMData,
+    getTooltipTextAFMData,
+} = require("../../data/geoChart/fixtures");
+
 module.exports = projectId => {
+    const locationData = getLocationAFMData(projectId);
+    const segmentData = getSegmentAFMData(projectId);
+    const tooltipTextData = getTooltipTextAFMData(projectId);
     return {
         executionResult: {
-            data: [["1005", "943", "179", "4726"], ["1005", "943", "179", "4726"]],
+            data: [SIZE_AFM_DATA.slice(0, 5), COLOR_AFM_DATA.slice(0, 5)],
             paging: {
-                count: [2, 50],
+                count: [2, 5],
                 offset: [0, 0],
-                total: [2, 50],
+                total: [2, 5],
             },
             headerItems: [
                 [
@@ -25,86 +36,7 @@ module.exports = projectId => {
                         },
                     ],
                 ],
-                [
-                    [
-                        {
-                            attributeHeaderItem: {
-                                name: "44.500000;-89.500000",
-                                uri: "/gdc/md/" + projectId + "/obj/694/elements?id=1808",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "39.000000;-80.500000",
-                                uri: "/gdc/md/" + projectId + "/obj/694/elements?id=1903",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "44.000000;-72.699997",
-                                uri: "/gdc/md/" + projectId + "/obj/694/elements?id=1870",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "31.000000;-100.000000",
-                                uri: "/gdc/md/" + projectId + "/obj/694/elements?id=1895",
-                            },
-                        },
-                    ],
-                    [
-                        {
-                            attributeHeaderItem: {
-                                name: "General Goods",
-                                uri: "/gdc/md/" + projectId + "/obj/790/elements?id=2027",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "General Goods",
-                                uri: "/gdc/md/" + projectId + "/obj/790/elements?id=2027",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "General Goods",
-                                uri: "/gdc/md/" + projectId + "/obj/790/elements?id=2027",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "General Goods",
-                                uri: "/gdc/md/" + projectId + "/obj/790/elements?id=2027",
-                            },
-                        },
-                    ],
-                    [
-                        {
-                            attributeHeaderItem: {
-                                name: "Wisconsin",
-                                uri: "/gdc/md/" + projectId + "/obj/694/elements?id=1808",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "West Virginia",
-                                uri: "/gdc/md/" + projectId + "/obj/694/elements?id=1903",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Vermont",
-                                uri: "/gdc/md/" + projectId + "/obj/694/elements?id=1870",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Texas",
-                                uri: "/gdc/md/" + projectId + "/obj/694/elements?id=1895",
-                            },
-                        },
-                    ],
-                ],
+                [locationData.slice(0, 5), segmentData.slice(0, 5), tooltipTextData.slice(0, 5)],
             ],
         },
     };
